# assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cloud CLI Auth API",
    "version": "1.0.0",
    "description": "Google-authenticated session and user-property API. Browser requests must include session cookies."
  },
  "servers": [
    {
      "url": "__HOSTNAME__"
    }
  ],
...
  "paths": {
...
    "/login": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Show the Google sign-in page",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "Optional URL to return to after sign-in",
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML sign-in page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Show the current user's profile page",
        "responses": {
          "200": {
            "description": "HTML profile page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      }
    },
...
    "/properties": {
      "get": {
        "tags": ["Properties"],
        "summary": "List the current user's properties",
        "responses": {
          "200": {
            "description": "Properties",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Property"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      },
      "put": {
        "tags": ["Properties"],
        "summary": "Create or update a property for the current user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved property",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Property"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "500": {
            "description": "Invalid input or storage failure"
          }
        }
      }
    },
    "/properties/{key}": {
      "parameters": [
        {
          "name": "key",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": ["Properties"],
        "summary": "Get a property for the current user",
        "responses": {
          "200": {
            "description": "Property",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Property"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "description": "Property not found"
          }
        }
      },
      "delete": {
        "tags": ["Properties"],
        "summary": "Delete a property for the current user",
        "responses": {
          "202": {
            "description": "Property deleted"
          },
          "400": {
            "description": "Missing property key"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "500": {
            "description": "Storage failure"
          }
        }
      }
    },
...
    "/session/token": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Issue a short-lived access token for an allowed audience",
        "description": "Requires an authenticated browser session and an Origin listed in AUTH_ALLOWED_ORIGINS.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["audience"],
                "properties": {
                  "audience": {
                    "type": "string",
                    "description": "A value configured in JWT_AUDIENCES"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bearer access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "400": {
            "description": "Invalid audience"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "description": "Origin is not allowed"
          },
          "503": {
            "description": "JWT service is not configured"
          }
        }
      }
    },
    "/authorize": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Start an OIDC authorization-code flow",
        "parameters": [
          { "name": "response_type", "in": "query", "required": true, "schema": { "type": "string", "const": "code" } },
          { "name": "client_id", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "redirect_uri", "in": "query", "required": true, "schema": { "type": "string", "format": "uri" } },
          { "name": "state", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "code_challenge", "in": "query", "required": true, "schema": { "type": "string" } },
          {
            "name": "code_challenge_method",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "const": "S256" }
          }
        ],
        "responses": {
          "302": { "description": "Redirect to sign-in or the registered callback with code and state" },
          "400": { "description": "Invalid or unregistered client request" }
        }
      }
    },
    "/token": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Exchange an OIDC authorization code",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": ["grant_type", "code", "client_id", "client_secret", "redirect_uri", "code_verifier"],
                "properties": {
                  "grant_type": { "type": "string", "const": "authorization_code" },
                  "code": { "type": "string" },
                  "client_id": { "type": "string" },
                  "client_secret": { "type": "string" },
                  "redirect_uri": { "type": "string", "format": "uri" },
                  "code_verifier": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access and identity tokens",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/OidcTokenResponse" } }
            }
          },
          "400": { "description": "Invalid request or authorization code" },
          "503": { "description": "JWT service is not configured" }
        }
      }
    },
    "/.well-known/jwks.json": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get public JWT signing keys",
        "responses": {
          "200": {
            "description": "JSON Web Key Set",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["keys"],
                  "properties": {
                    "keys": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "JWT service is not configured"
          }
        }
      }
    },
    "/userinfo": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get a user's profile from a verified JWT",
        "parameters": [
          {
            "name": "X-Auth-Audience",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user's profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or incorrectly-audienced bearer token"
          },
          "404": {
            "description": "User no longer exists"
          }
        }
      }
    },
...
    "/webauthn/register/options": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Create passkey registration options for the current user",
        "responses": {
          "200": { "description": "WebAuthn registration options" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/webauthn/register/verify": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a passkey for the current user",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": {
          "201": { "description": "Passkey registered" },
          "400": { "description": "Invalid WebAuthn response" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/webauthn/authentication/options": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Create passkey authentication options",
        "parameters": [{ "name": "login_hint", "in": "query", "required": false, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "WebAuthn authentication options" } }
      }
    },
    "/webauthn/authentication/verify": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate with a passkey and create a session",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": {
          "204": { "description": "Authenticated" },
          "401": { "description": "Invalid or revoked passkey" }
        }
      }
    },
    "/webauthn/credentials": {
      "get": {
        "tags": ["Authentication"],
        "summary": "List the current user's passkeys",
        "responses": {
          "200": {
            "description": "Registered passkeys",
            "content": { "application/json": { "schema": { "type": "array" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/webauthn/credentials/{credentialId}": {
      "delete": {
        "tags": ["Authentication"],
        "summary": "Revoke one of the current user's passkeys",
        "parameters": [{ "name": "credentialId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Passkey revoked" },
          "404": { "description": "Passkey not found" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/recovery-codes": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Generate a new set of one-time recovery codes",
        "responses": {
          "200": {
            "description": "Recovery codes, shown once",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "codes": { "type": "array", "items": { "type": "string" } } }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/qr-login/start": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Create a QR login approval transaction",
        "parameters": [{ "name": "url", "in": "query", "required": false, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Transaction token, QR data URL, and PWA URL" } }
      }
    },
    "/qr-login/status": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Poll a laptop QR login transaction",
        "parameters": [{ "name": "transaction", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Pending, denied, or approved transaction" },
          "410": { "description": "Expired transaction" }
        }
      }
    },
    "/qr-login/details": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Read a QR login transaction from an authenticated phone",
        "parameters": [{ "name": "transaction", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Login details" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "410": { "description": "Expired transaction" }
        }
      }
    },
    "/qr-login/approve": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Approve a QR login from the authenticated phone",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["transaction"],
                "properties": { "transaction": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Approved" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "410": { "description": "Expired transaction" }
        }
      }
    },
    "/qr-login/deny": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Deny a QR login from the authenticated phone",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["transaction"],
                "properties": { "transaction": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Denied" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "410": { "description": "Expired transaction" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": ["id", "name", "email", "photo"],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "photo": {
            "type": "string"
          }
        }
      },
      "PropertyInput": {
        "type": "object",
        "required": ["key", "value"],
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "Property": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PropertyInput"
          },
          {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string"
              },
              "userId": {
                "type": "string"
              }
            }
          }
        ]
      },
      "AccessToken": {
        "type": "object",
        "required": ["access_token", "token_type", "expires_in"],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "const": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 60,
            "maximum": 900
          }
        }
      },
      "OidcTokenResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/AccessToken" },
          {
            "type": "object",
            "required": ["id_token"],
            "properties": {
              "id_token": { "type": "string" }
            }
          }
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthenticated": {
        "description": "No authenticated session"
      }
    }
  }
}